Added redux, integrated with form
