+++ reactWidget/package.json
@@ -22,6 +22,8 @@
     "babel-preset-stage-0": "^6.5.0",
     "react": "^0.14.7",
     "react-dom": "^0.14.7",
+    "react-redux": "^5.0.1",
+    "redux": "^3.6.0",
     "validatorjs": "^2.0.5",
     "webpack": "^1.12.14",
     "webpack-dev-server": "^1.14.1"
