+++ reactWidget/src/client/app/BirthdayComponent.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import ReactDom from 'react-dom';
 import {getApiCall, postApiCall} from './Api.js';
+import {connect} from 'react-redux'
+import {addEmployee} from './actions/actions'
 
 require('../sass/styles.global.scss');
 
@@ -9,14 +11,8 @@ const TOTAL_INPUTS = 7;
 const BirthdayComponent = React.createClass({
 	postData(e) {
 		e.preventDefault();
-		const successCallback = (value) => {
-		  this.resetForm();
-		  this.gotReponseFromServer(true);
-		}
-		const errorCallBack = (value) => {
-		  this.gotReponseFromServer(false);
-		}
-		postApiCall('/RewardsWidgetAPI/employees', this.getDataToSend(), successCallback, errorCallBack);
+		this.props.addEmployee(this.getDataToSend());
+		this.gotReponseFromServer(true);
 	},
 	getDataToSend() {
 		return ({
@@ -88,6 +84,10 @@ const BirthdayComponent = React.createClass({
 	},
   	render() {
 		this.progressBarIncrement();
+
+		console.log(this.props)
+
+
 		var formMessage = this.state.gotResponse ? this.state.isSuccess ? "success" : "error" : "";
 		formMessage = this.state.isSubmitting ? "loading" : formMessage;
     	return (
@@ -147,4 +147,24 @@ const BirthdayComponent = React.createClass({
  	}
 })
 
-export default BirthdayComponent;
+const mapStateToProps = (state) => {
+  return {
+    employees: state.employees
+  }
+}
+
+const mapDispatchToProps = (dispatch) => {
+  return {
+    addEmployee: (employee) => {
+      dispatch(addEmployee(employee))
+    }
+  }
+}
+
+const AddBirthdayComponent = connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(BirthdayComponent)
+
+export default AddBirthdayComponent
+
