+++ reactWidget/src/client/app/actions/actions.js
@@ -1,0 +1,11 @@
+let nextTodoId = 0
+
+export const addEmployee = (employee) => {
+  return {
+    type: 'ADD_EMPLOYEE',
+    employee: {
+      id: nextTodoId++,
+      employee
+    }
+  }
+}
