+++ reactWidget/src/client/app/reducers/employeeReducer.js
@@ -1,0 +1,13 @@
+const employeeReducer = (state = [], action) => {
+  switch (action.type) {
+    case 'ADD_EMPLOYEE':
+      return [
+      ...state,
+      action.employee,
+      ];
+    default:
+      return state
+  }
+}
+
+export default employeeReducer
