+++ reactWidget/src/client/app/reducers/index.js
@@ -1,0 +1,8 @@
+import {combineReducers} from 'redux'
+import employeeReducer from './employeeReducer'
+
+const combinedReducer = combineReducers({
+  employees: employeeReducer
+})
+
+export default combinedReducer
